Add sequences and negative numbers
In this commit I add sequences (that is, the ability for the root level (and
later the insides of a function call) to be a sequence of one or more comma
separated expressions), and the ability to actually write negative numbers (you
would previously have had to have had use "0-x" to get -x). I also added the
ability to negate any term along side that, so you can also just put a negative
in front of a bracketed expression, for instance.