Setup structure for functions
This commit makes a reasonably substantial restrucutring of the node system,
having individual nodes yeild arrays of number, and sequence nodes yeild arrays
or arrays of numbers, to allow functions to still take and return multiple
things, but without forcing me to implement some kind of infinite everything-
is-an-array recursion.

It also adds the function node, and new tests for functions and sequences, and
updates the tests to expect this new array or arrays structure.