Update grammar and add typescript parse table
Whilst actually trying to make a parse table I discovered a bunch of changes,
and, indeed, fixes, to the grammar. Unfortunately this means all the working
isn't exactly accurate to what would be required to get the finished product,
but I'm pretty sure that the actual values in the table are right.