Remove a bunch of stuff that is no longer relevant
This was a lot of stuff involved in the recursive descent parse, that is going
to be replaced by the table parse. Some of it will be replaced, I will, for
instance, still need something like these nodes, but they'll probably be
re-written.

The parser itself is also completely going, of course.

I only got rid of the exceptions because they had dashes in the file names,
they'll probably be replaced exactly.