Add source and terminal enums
These really should have gone along with the grammar and the parse table.
Whoops.

If you can't tell, I'm commiting all of this at the same time, just breaking it
up in the hope of it being slightly more logical. I held of commiting anything
for a while because I didn't want to break master, which really is ridiculous,
because that is what branches are for.