Add Sequence interface
This is the typing solution I was looking for but didn't find last time. A
Sequence is an array. An element of a Sequence is either a number, or, wait for
it, a Sequence.

I didn't know how to specify this in typescript for ages, turns out it's really
straight forward. This gives me my arbitrarially nested arrays that I feel it
makes sense for SweetRolls code to evaluate to.